Added DisplaySystem
magma::DisplaySystem is a class to help with sawpchain, recreation and handling userData associated with frames.
It automaticly (re)calls constructors of the different userdata provided.
It is undocumented for the moment, and this design isn't final, this pull request is mostly for reviews.

Also:
- made magma::Instance movable
- fixed magma::CommandPool destructor (ressources weren't released)
- made surface be a handletype (breaking change)
- moved some includes (with the goal to make files more self-contained)

These other changes are rather minor, but where needed.
Please do not merge his pull request before monday, I want to see how it works on https://github.com/raven-os/feathers first and refine it.
Reviews and ideas are welcome.