Added `makeCreateInfo` and `makeCreateInfoNoFlags`
This takes advantage of function partial specialization to simplify the API.
Code was tested using wasted_prophecies and is confirmed to work.
I'd be for deprecating usage of the struct above, and removing or moving it to
to magma::impl once feathers no longer depends on it.

Some documentation was also added for the 2 new functions.